requirements.txt
fastapi
uvicorn
pydantic
requests
msgpack
pycryptodome
slowapi
python-dotenv
python-crontab